feat(beta/skills): add SkillPlugin and reorganize skill tests
Move the SkillPlugin factory into the framework. It builds a Plugin that
injects the local-skill catalog (name + description) into the system
prompt on every turn and exposes load_skill / run_skill_script tools —
following the Skills spec (metadata on load) instead of a list_skills
tool call. Exposed lazily from autogen.beta.tools.skills to avoid the
agent -> tools -> skills import cycle.

Split the skills test suite into test/beta/tools/skills/ (loader,
toolkit, plugin, skill_search) with a shared skill_tree fixture. Plugin
tests use TestConfig/TrackingConfig and assert via reply.context.prompt
and tracked tool-result messages per test/beta/CLAUDE.md.
