+++ .gitignore
@@ -1,5 +1,6 @@
 .agents
 .claude
+.codegraph
 .rtk
 skills-lock.json
